requirements.txt
GitPython==3.1.43
PyYAML==6.0.2
scikit-learn==1.5.1
joblib==1.4.2

